Update 8. App Exercise - Mile Times and Congratulations
In the Steps struct below, add a willSet to the steps property that will check if the new value is equal to goal, and if it is, prints a congratulatory message. Create an instance of Steps where steps is 9999 and goal is 10000, then call takeStep() and see if your message is printed to the console.