Update 10. App Exercise - Type Properties and Methods
It may be helpful to have a few type properties on RunningWorkout representing unit conversions (i.e. meters to mile, feet to meters, etc.). Go back and add a type property for meterInFeet and assign it 3.28084. Then add a type property for mileInMeters and assign it 1600.0. Print both of these values below.