Update 1. Exercise - Define a Base Class
Create a Spaceship class with three variable properties: name, health, and position. The default value of name should be an empty string and health should be 0. position will be represented by an Int where negative numbers place the ship further to the left and positive numbers place the ship further to the right. The default value of position should be 0.