Update 1. Exercise - Define a Base Class
Go back and add a method called moveLeft() to the definition of Spaceship. This method should adjust the position of the spaceship to the left by one. Add a similar method called moveRight() that moves the spaceship to the right. Once these methods exist, use them to move falcon to the left twice and to the right once. Print the new position of falcon after each change in position.